Replaced old staff system detection

+++ moonshine/structure/staves.py
@@ -42,7 +42,7 @@ def staves(page):
     staves = hough.hough_paths(lines)
     # Filter out staves which are too short
     good_staves = (staves[:, 1] - staves[:, 0]) > page.orig_size[1] / 2.0
-    page.staves = staves[good_staves]
+    page.staves = staves[good_staves].astype(np.int32)
     return page.staves
 
 def show_staff_segments(page):
